[FIR] Use member from purely implemented supertype for result of intersection
Java classes that are annotated with @PurelyImplements have supertypes
without platform types in their generic arguments.
However, we don't necessarily return those platform-type-free members
from the scope of such a Java class because of the complex rules
of intersection overrides.
In this commit, an API is introduced to intercept the computation of
intersection overrides.
This API is used in JavaClassUseSiteMemberScope so that if the group
of overridden members contains one from the putely implemented
supertype, this member is always selected as the most specific one
and its signature ends up being used.

This fixes some false-negative nullability errors when calling methods
of Java collections.

#KT-59921 Fixed